implement usleep(...) workaround for windows
Windows doesn't provide a (working) usleep(...) function. In cross-
compile environment the usleep(...) function is a almost zero-wait
function and therefore absolutely unusable.

For this we insert a abstraction layer "osi.h".
Linux build is unaffected from this change, there is still the syscall
"usleep" used.
For Windows we use the "Sleep(..)" syscall for times >= 1000us and all
below we querry the performance counter and do a busy-wait.

Signed-off-by: Hannes Schmelzer <hannes.schmelzer@br-automation.com>